BE-57: added security layer, feign client implemented
