+++ build.gradle
@@ -36,7 +36,7 @@ checkstyle {
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    //implementation 'org.springframework.boot:spring-boot-starter-security'
+    implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
     compile group: 'io.jsonwebtoken', name: 'jjwt', version: '0.9.1'
     compileOnly 'org.projectlombok:lombok'
@@ -52,7 +52,7 @@ dependencies {
     testCompile platform("org.spockframework:spock-bom:2.0-M1-groovy-2.5")
     testCompile "org.spockframework:spock-core"
 
-    //testImplementation 'org.springframework.security:spring-security-test'
+    testImplementation 'org.springframework.security:spring-security-test'
 
 }
