--- src/main/java/az/gdg/article/MsArticleApplication.java
@@ -1,15 +1,0 @@
-package az.gdg.article;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.mongodb.config.EnableMongoAuditing;
-
-@EnableMongoAuditing
-@SpringBootApplication
-public class MsArticleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MsArticleApplication.class, args);
-    }
-
-}
