--- src/main/java/az/gdg/article/model/ArticleRequest.java
@@ -1,23 +1,0 @@
-package az.gdg.article.model;
-
-import az.gdg.article.model.dto.TagDTO;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ArticleRequest {
-
-    private String title;
-    private String content;
-    private Integer clapCount;
-    private Integer readCount;
-    private List<TagDTO> tags;
-    private List<CommentRequest> comments;
-}
