--- src/main/java/az/gdg/article/model/TagRequest.java
@@ -1,5 +1,0 @@
-package az.gdg.article.model.dto;
-
-public class TagRequest {
-    private String name;
-}
