--- src/main/java/az/gdg/article/model/dto/TagDTO.java
@@ -1,14 +1,0 @@
-package az.gdg.article.model.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Data
-public class TagDTO {
-    private String name;
-}
