--- src/main/java/az/gdg/article/model/entity/TagEntity.java
@@ -1,19 +1,0 @@
-package az.gdg.article.model.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Document(collection = "tags")
-public class TagEntity {
-    @Id
-    private String id;
-    private String name;
-}
