--- src/main/java/az/gdg/article/repository/TagRepository.java
@@ -1,9 +1,0 @@
-package az.gdg.article.repository;
-
-import az.gdg.article.model.entity.TagEntity;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TagRepository extends MongoRepository<TagEntity, String> {
-}
