--- src/main/java/az/gdg/article/service/ArticleService.java
@@ -1,9 +1,0 @@
-package az.gdg.article.service;
-
-import az.gdg.article.model.ArticleRequest;
-
-public interface ArticleService {
-
-    String addDraft(String token, ArticleRequest articleRequest);
-
-}
