+++ src/main/java/az/gdg/msarticle/exception/InvalidTokenException.java
@@ -1,0 +1,7 @@
+package az.gdg.msarticle.exception;
+
+public class InvalidTokenException extends RuntimeException {
+    public InvalidTokenException(String message) {
+        super(message);
+    }
+}
