+++ src/main/java/az/gdg/msarticle/model/TagRequest.java
@@ -1,0 +1,5 @@
+package az.gdg.msarticle.model.dto;
+
+public class TagRequest {
+    private String name;
+}
