+++ src/main/java/az/gdg/msarticle/model/client/auth/HttpHeader.java
@@ -1,0 +1,9 @@
+package az.gdg.msarticle.model.client.auth;
+
+public abstract class HttpHeader {
+    public static final String X_AUTH_TOKEN = "X-Auth-Token";
+
+    private HttpHeader() {
+
+    }
+}
