+++ src/main/java/az/gdg/msarticle/model/client/auth/UserInfo.java
@@ -1,0 +1,14 @@
+package az.gdg.msarticle.model.client.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class UserInfo {
+    private String userId;
+}
