+++ src/main/java/az/gdg/msarticle/model/dto/TagDTO.java
@@ -1,0 +1,14 @@
+package az.gdg.msarticle.model.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class TagDTO {
+    private String name;
+}
