+++ src/main/java/az/gdg/msarticle/repository/ArticleRepository.java
@@ -1,0 +1,10 @@
+package az.gdg.msarticle.repository;
+
+import az.gdg.msarticle.model.entity.ArticleEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ArticleRepository extends MongoRepository<ArticleEntity, String> {
+
+}
