+++ src/main/java/az/gdg/msarticle/repository/TagRepository.java
@@ -1,0 +1,9 @@
+package az.gdg.msarticle.repository;
+
+import az.gdg.msarticle.model.entity.TagEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TagRepository extends MongoRepository<TagEntity, String> {
+}
