+++ src/main/java/az/gdg/msarticle/security/UserAuthentication.java
@@ -1,0 +1,57 @@
+package az.gdg.msarticle.security;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+
+import java.util.Collection;
+
+public class UserAuthentication implements Authentication {
+
+    private final String userId;
+    private boolean authenticated;
+
+
+    public UserAuthentication(String userId, boolean authenticated) {
+        this.userId = userId;
+        this.authenticated = authenticated;
+    }
+
+    public String getUserId() {
+        return userId;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @Override
+    public Object getCredentials() {
+        return null;
+    }
+
+    @Override
+    public Object getDetails() {
+        return null;
+    }
+
+    @Override
+    public String getPrincipal() {
+        return getUserId();
+    }
+
+    @Override
+    public boolean isAuthenticated() {
+        return authenticated;
+    }
+
+    @Override
+    public void setAuthenticated(boolean authenticated) {
+        this.authenticated = authenticated;
+    }
+
+    @Override
+    public String getName() {
+        return null;
+    }
+}
