+++ src/main/java/az/gdg/msarticle/service/ArticleService.java
@@ -1,0 +1,9 @@
+package az.gdg.msarticle.service;
+
+import az.gdg.msarticle.model.ArticleRequest;
+
+public interface ArticleService {
+
+    String addDraft(String token, ArticleRequest articleRequest);
+
+}
