+++ src/main/resources/application.yml
@@ -10,10 +10,6 @@ spring:
     mongodb:
       uri: ${MONGODB_URI}
 
-jwt:
-  secret: ${SECRET_KEY}
-  expiration: 9000
-  header: X-Auth-Token
 swagger:
   enabled: true
