--- src/test/groovy/az/gdg/article/MsArticleApplicationTest.groovy
@@ -1,6 +1,0 @@
-package az.gdg.article
-
-import spock.lang.Specification
-
-class MsArticleApplicationTest extends Specification {
-}
