+++ src/test/groovy/az/gdg/msarticle/MsArticleApplicationTest.groovy
@@ -1,0 +1,6 @@
+package az.gdg.msarticle
+
+import spock.lang.Specification
+
+class MsArticleApplicationTest extends Specification {
+}
